feat(payment-stripe): add calculation service
